+++ packages/template-browser/src/importer.ts
@@ -1,0 +1,8 @@
+/** Recommended headless template-import workflow contract. */
+export * from "../../../src/template-package/import/templateImportValidation";
+export * from "../../../src/template-package/import/templateImportIntegrity";
+export * from "../../../src/template-package/import/templateImportWizard";
+export {
+  formatRequiredFontFace,
+  type ExactFontSetupErrorCode,
+} from "../../../src/template-package/fonts/exactFontSetup";